Refactored the whole bot to Classes, cleanups and small fixes.
finance_helper.py -> changed default_start and default_end to something more recent

helpers.py -> created a file for all helper functions which are used on different places in the code

commands.py -> File for all unsorted commands

greetings.py -> File for the greeting commands
greetings_cog.py -> removed

stocks.py -> File for the stock commands

tweets.py -> File for all tweet related commands

created a test folder which could contain all coming test files including unit tests.

removed bot_cmd.py and separated the functions into multiple files.

added main.py for basic bot features.

moved everything into classes to make it more clear and modular

changed requirements.txt so it will not use highly outdated libraries

updated ALL dependencies
