+++ modules/greetings.py
@@ -22,3 +22,7 @@ class Greetings(commands.Cog):
         else:
             await ctx.send('Hello {0.name}... This feels familiar.'.format(member))
         self._last_member = member
+
+
+def setup(bot):
+    bot.add_cog(Greetings(bot))
