+++ requirements.txt
@@ -1,6 +1,6 @@
-discord.py==1.3.4
-yfinance==0.1.54
-lxml==4.6.2
-numpy==1.19.3
-matplotlib==3.3.3
-searchtweets==1.7.6
+discord.py>=1.3.4
+yfinance>=0.1.54
+lxml>=4.6.2
+numpy>=1.19.3
+matplotlib>=3.3.3
+searchtweets>=1.7.6
